pyproject.toml
[tool.poetry]
name = "kgx"
version = "0.0.0"
description = "A Python library and set of command line utilities for exchanging Knowledge Graphs (KGs) that conform to or are aligned to the Biolink Model."
authors = ["Deepak Unni <deepak.unni3@gmail.com>", "Richard Bruskiewich <richard.bruskiewich@delphinai.com>", "Sierra Moxon <smoxon@lbl.gov>"]

license = "BSD"
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3"
]

packages = [
    { include = "kgx" }
]

[tool.poetry.dependencies]
python = "^3.9"
Click = "*"
SPARQLWrapper = ">=1.8.2"
bmt = "^1.4.6"
cachetools = "^5.0.0"
deprecation = "^2.1.0"
docker = "^7.0.0"
docutils = ">=0.18.1"
duckdb = "^1.0.0"
exceptiongroup = {version = "^1.0.0", python = "<3.11"}
ijson = "^3.1.3"
inflection = "^0.5.1"
jsonlines = "^4.0.0"
jsonstreams = "^0.6.0"
linkml = "^1.9.1"
linkml-runtime = "^1.9.1"
linkml-map = "^0.3.8"
mypy = "*"
neo4j = ">=4.4.10"
python-arango = ">=7.0.0"
networkx = "*"
ordered-set = "^4.0.2"
pandas = ">1.0.3"
prefixcommons = "^0.1.4"
prologterms = "^0.0.6"
pytest = "*"
python-dateutil = "^2.8.1"
pyyaml = "*"
rdflib = ">=6.0.0"
pyjelly = "^0.6.2"
shexjsg = "*"
stringcase = "^1.2.0"
terminaltables = "^3.1.0"
tox = "^4.13.0"
validators = "^0.20.0"
pyarrow = "^18.0.0"
recommonmark = "0.5.0"
sphinx = "5.3.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
furo = {version = "^2023.03.27", extras = ["docs"]}
sphinxcontrib-mermaid = {version = "^0.7.1", extras = ["docs"]}
sphinx = "*"
sphinx-click = "*"
sphinx-rtd-theme = "*"
myst-parser = "*"
myst-nb = "*"  # <-- Add this line
matplotlib = ">=3.7"
sphinx-jinja = "^2.0.2"
sphinxcontrib-programoutput = "^0.17"
sphinx-design = "*"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[build-system]

requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
kgx = "kgx.cli:cli"
